Move request_with_user_and_mock_crate into team tests
This is the only place this function is used for now; I want to get rid
of it soon but I just want to contain it for the moment.
